# BE-Challenge/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "API for user authentication, authorization, and management"
  },
  "paths": {
...
    "/user/{id}": {
      "get": {
        "summary": "Get user by ID",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to fetch",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "123"
                    },
                    "name": {
                      "type": "string",
                      "example": "John Doe"
                    },
                    "email": {
                      "type": "string",
                      "example": "john.doe@example.com"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/users/{id}": {
      "patch": {
        "summary": "Update user by ID",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/user/password/{id}": {
      "patch": {
        "summary": "Update user password",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to update password",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated successfully"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/user/updateAccountType/{id}": {
      "put": {
        "summary": "Update user account type",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to update account type",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "accountType": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User account type updated successfully"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/challenge": {
      "get": {
        "summary": "Get all challenges",
        "tags": [
          "Challenge"
        ],
        "responses": {
          "200": {
            "description": "A list of challenges",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "123"
                      },
                      "title": {
                        "type": "string",
                        "example": "Challenge 1"
                      },
                      "description": {
                        "type": "string",
                        "example": "Description of challenge 1"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new challenge",
        "tags": [
          "Challenge"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                },
                "required": [
                  "title",
                  "description"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Challenge created successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/challenge/{id}": {
      "get": {
        "summary": "Get challenge by ID",
        "tags": [
          "Challenge"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the challenge to fetch",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single challenge",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "123"
                    },
                    "title": {
                      "type": "string",
                      "example": "Challenge 1"
                    },
                    "description": {
                      "type": "string",
                      "example": "Description of challenge 1"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Challenge not found"
          }
        }
      },
      "put": {
        "summary": "Update challenge by ID",
        "tags": [
          "Challenge"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the challenge to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Challenge updated successfully"
          },
          "404": {
            "description": "Challenge not found"
          }
        }
      },
      "delete": {
        "summary": "Delete challenge by ID",
        "tags": [
          "Challenge"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the challenge to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Challenge deleted successfully"
          },
          "404": {
            "description": "Challenge not found"
          }
        }
      }
    },
...
    "/challengeUser/user/{id}": {
      "get": {
        "summary": "Get all challenges for a user",
        "tags": [
          "ChallengeUser"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to fetch challenges for",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of user challenges",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "123"
                      },
                      "userID": {
                        "type": "string",
                        "example": "user123"
                      },
                      "challengeID": {
                        "type": "string",
                        "example": "challenge123"
                      },
                      "status": {
                        "type": "string",
                        "example": "completed"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/challengeUser": {
      "post": {
        "summary": "Create a new user challenge",
        "tags": [
          "ChallengeUser"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userID": {
                    "type": "string"
                  },
                  "challengeID": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                },
                "required": [
                  "userID",
                  "challengeID",
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User challenge created successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/challengeUser/{userID}/{challengeID}": {
      "get": {
        "summary": "Get a user challenge by user ID and challenge ID",
        "tags": [
          "ChallengeUser"
        ],
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "required": true,
            "description": "ID of the user",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "challengeID",
            "in": "path",
            "required": true,
            "description": "ID of the challenge",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single user challenge",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "123"
                    },
                    "userID": {
                      "type": "string",
                      "example": "user123"
                    },
                    "challengeID": {
                      "type": "string",
                      "example": "challenge123"
                    },
                    "status": {
                      "type": "string",
                      "example": "completed"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User challenge not found"
          }
        }
      },
      "put": {
        "summary": "Update a user challenge",
        "tags": [
          "ChallengeUser"
        ],
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "required": true,
            "description": "ID of the user",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "challengeID",
            "in": "path",
            "required": true,
            "description": "ID of the challenge",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  },
                  "contentImg": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User challenge updated successfully"
          },
          "404": {
            "description": "User challenge not found"
          }
        }
      }
    },
    "/challengeUser/{userID}": {
      "get": {
        "summary": "Get all challenges for a user",
        "tags": [
          "ChallengeUser"
        ],
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "required": true,
            "description": "ID of the user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of user challenges",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "123"
                      },
                      "userID": {
                        "type": "string",
                        "example": "user123"
                      },
                      "challengeID": {
                        "type": "string",
                        "example": "challenge123"
                      },
                      "status": {
                        "type": "string",
                        "example": "completed"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}